feat: Relation between User and Transactions and remove unused Transaction Controller
