+++ src/transactions/entitites/transaction.entity.ts
@@ -1,0 +1,23 @@
+import { User } from 'src/users/entities/user.entity';
+import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
+
+@Entity({ name: 'transactions' })
+export class Transaction {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column({ nullable: false })
+  type: string;
+
+  @Column({ nullable: false })
+  description: string;
+
+  @Column({ type: 'datetime', default: () => 'CURRENT_TIMESTAMP' })
+  createdAt: Date;
+
+  @Column({ nullable: false })
+  userId: number;
+
+  @ManyToOne(() => User, (user) => user.transactions)
+  user: User;
+}
