--- src/transactions/transactions.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('transactions')
-export class TransactionsController {}
