+++ src/transactions/transactions.module.ts
@@ -1,9 +1,7 @@
 import { Module } from '@nestjs/common';
-import { TransactionsController } from './transactions.controller';
 import { TransactionsService } from './transactions.service';
 
 @Module({
-  controllers: [TransactionsController],
   providers: [TransactionsService],
 })
 export class TransactionsModule {}
