+++ src/users/entities/user.entity.ts
@@ -1,4 +1,5 @@
 import { Task } from 'src/tasks/entities/task.entity';
+import { Transaction } from 'src/transactions/entitites/transaction.entity';
 import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity({ name: 'users' })
@@ -17,4 +18,7 @@ export class User {
 
   @OneToMany(() => Task, (task) => task.user)
   tasks: Task[];
+
+  @OneToMany(() => Transaction, (transaction) => transaction.user)
+  transactions: Transaction[];
 }
